feat : 엔티티 클래스 추가, 참조 관계에 cascade 및 fetch.lazy 적용
