--- src/main/java/com/practice/survey/domain/entity/ResponseItem.java
@@ -1,46 +1,0 @@
-package com.practice.survey.domain.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import org.hibernate.annotations.CreationTimestamp;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-@Entity
-public class ResponseItem extends BaseTime{
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long resposeItemId;
-
-    @CreationTimestamp
-    @Column(name="submitted_at",nullable = false,updatable = false)
-    private LocalDateTime submittedAt;
-
-    @Column(columnDefinition = "TEXT")
-    private String responseValue ;
-
-    @ManyToOne
-    @JoinColumn(name = "response_id")
-    private Response response;
-
-    @ManyToOne
-    @JoinColumn(name = "item_id")
-    private SurveyItem surveyItem;
-
-}
