--- src/main/java/com/practice/survey/domain/entity/SurveyVersion.java
@@ -1,36 +1,0 @@
-package com.practice.survey.domain.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access= AccessLevel.PROTECTED)
-@ToString
-@Entity
-public class SurveyVersion extends BaseTime {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long versionId;
-
-    @Column(nullable = false)
-    private int versionNumber;
-
-    @ManyToOne
-    @JoinColumn(name = "survey_id")
-    private Survey survey;
-
-}
