+++ src/main/java/com/practice/survey/surveyItem/model/entity/SurveyItem.java
@@ -1,16 +1,9 @@
-package com.practice.survey.domain.entity;
+package com.practice.survey.surveyItem.model.entity;
 
-import com.practice.survey.domain.enums.InputType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
+import com.practice.survey.common.model.entity.BaseTime;
+import com.practice.survey.surveyVersion.model.entity.SurveyVersion;
+import com.practice.survey.surveyItem.model.enums.InputType;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -20,10 +13,10 @@ import lombok.ToString;
 @Getter
 @Builder
 @AllArgsConstructor
-@NoArgsConstructor(access= AccessLevel.PROTECTED)
+@NoArgsConstructor
 @ToString
 @Entity
-public class SurveyItem extends BaseTime{
+public class SurveyItem extends BaseTime {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -42,10 +35,11 @@ public class SurveyItem extends BaseTime{
     @Column(name = "input_type",nullable = false)
     private InputType inputType;
 
+    @Builder.Default
     @Column(nullable = false)
-    private boolean isRequired=true;
+    private boolean isRequired = true;
 
-    @ManyToOne
-    @JoinColumn(name = "version_id")
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    @JoinColumn(name = "version_id", nullable = false)
     private SurveyVersion version;
 }
